updated code to fix webpack build failure in cypress tests
