+++ public/utils/constants.ts
@@ -3,14 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { CoreStart, IUiSettingsClient, SimpleSavedObject } from 'opensearch-dashboards/public';
+import { SimpleSavedObject } from 'opensearch-dashboards/public';
 import { Detector, LogType, ServerResponse } from '../../types';
 import { DetectorInput, PeriodSchedule } from '../../models/interfaces';
 import { DetectorHit } from '../../server/models/interfaces';
 import _ from 'lodash';
 import { euiPaletteColorBlind } from '@elastic/eui';
-import { createGetterSetter } from '../../../../src/plugins/opensearch_dashboards_utils/common';
-import { NavigationPublicPluginStart } from '../../../../src/plugins/navigation/public';
 
 export const DATE_MATH_FORMAT = 'YYYY-MM-DDTHH:mm:ss.SSSZ';
 export const MAX_RECENTLY_USED_TIME_RANGES = 5;
@@ -243,21 +241,3 @@ export enum AlertTabId {
   ThreatIntel = 'threat-intel',
   Correlations = 'correlations',
 }
-
-export const [getUISettings, setUISettings] = createGetterSetter<IUiSettingsClient>('UISettings');
-
-export const getUseUpdatedUx = () => {
-  return getUISettings().get('home:useNewHomePage', false);
-};
-
-export const [getNavigationUI, setNavigationUI] = createGetterSetter<
-  NavigationPublicPluginStart['ui']
->('navigation');
-
-export const [getApplication, setApplication] = createGetterSetter<CoreStart['application']>(
-  'application'
-);
-
-export const [getBreadCrumbsSetter, setBreadCrumbsSetter] = createGetterSetter<
-  CoreStart['chrome']['setBreadcrumbs']
->('breadCrumbSetter');
